# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leadership Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  manager: "manager",
  participant: "participant",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
}

export interface LoginBody {
  userId: number;
}

export type CaseStatus = (typeof CaseStatus)[keyof typeof CaseStatus];

export const CaseStatus = {
  draft: "draft",
  self_assessment_open: "self_assessment_open",
  external_assessment_open: "external_assessment_open",
  collecting_responses: "collecting_responses",
  ready_for_report: "ready_for_report",
  report_generated: "report_generated",
  reviewed: "reviewed",
  released: "released",
  closed: "closed",
} as const;

export interface Case {
  id: number;
  title: string;
  targetPersonId: number;
  targetPersonName: string;
  status: CaseStatus;
  createdById: number;
  createdByName: string;
  createdAt: string;
  updatedAt: string;
  respondentCount: number;
  completedRespondentCount: number;
  /** @nullable */
  selfAssessmentStatus: string | null;
}

export type AssessmentInstanceType =
  (typeof AssessmentInstanceType)[keyof typeof AssessmentInstanceType];

export const AssessmentInstanceType = {
  self: "self",
  external: "external",
} as const;

export type AssessmentInstanceStatus =
  (typeof AssessmentInstanceStatus)[keyof typeof AssessmentInstanceStatus];

export const AssessmentInstanceStatus = {
  pending: "pending",
  in_progress: "in_progress",
  submitted: "submitted",
} as const;

export interface AssessmentInstance {
  id: number;
  caseId: number;
  type: AssessmentInstanceType;
  status: AssessmentInstanceStatus;
  /** @nullable */
  respondentName: string | null;
  /** @nullable */
  respondentEmail: string | null;
  /** @nullable */
  token: string | null;
  /** @nullable */
  submittedAt: string | null;
  createdAt: string;
}

export type ReportType = (typeof ReportType)[keyof typeof ReportType];

export const ReportType = {
  self: "self",
  external: "external",
  comparison: "comparison",
} as const;

export type ReportReleaseState =
  (typeof ReportReleaseState)[keyof typeof ReportReleaseState];

export const ReportReleaseState = {
  draft: "draft",
  reviewed: "reviewed",
  released: "released",
} as const;

export type ReportData = { [key: string]: unknown };

export interface Report {
  id: number;
  caseId: number;
  type: ReportType;
  generatedAt: string;
  releaseState: ReportReleaseState;
  data: ReportData;
}

export interface CaseDetail {
  id: number;
  title: string;
  targetPersonId: number;
  targetPersonName: string;
  status: string;
  createdById: number;
  createdByName: string;
  createdAt: string;
  updatedAt: string;
  assessments: AssessmentInstance[];
  reports: Report[];
}

export interface CreateCaseBody {
  title: string;
  targetPersonId: number;
  includeSelfAssessment?: boolean;
}

export interface UpdateCaseBody {
  status?: string;
  title?: string;
}

export interface AddRespondentBody {
  name: string;
  email: string;
}

export type ResponseSetAnswers = { [key: string]: number | null };

export type ResponseSetComments = { [key: string]: string };

export interface ResponseSet {
  assessmentInstanceId: number;
  answers: ResponseSetAnswers;
  comments: ResponseSetComments;
  /** @nullable */
  savedAt: string | null;
}

export type SaveAnswersBodyAnswers = { [key: string]: number | null };

export type SaveAnswersBodyComments = { [key: string]: string };

export interface SaveAnswersBody {
  answers: SaveAnswersBodyAnswers;
  comments: SaveAnswersBodyComments;
}

export type QuestionScaleLabels = { [key: string]: string };

export interface Question {
  id: string;
  text: string;
  scaleMin: number;
  scaleMax: number;
  scaleLabels: QuestionScaleLabels;
  allowComment: boolean;
}

export interface QuestionSection {
  id: string;
  title: string;
  description: string;
  questions: Question[];
}

export interface QuestionnaireStructure {
  sections: QuestionSection[];
}

export interface DashboardSummary {
  totalCases: number;
  activeCases: number;
  completedCases: number;
  pendingReports: number;
  totalRespondents: number;
  completedAssessments: number;
}

export interface CaseWithProgress {
  id: number;
  title: string;
  status: string;
  /** @nullable */
  selfAssessmentStatus: string | null;
  /** @nullable */
  selfAssessmentId: number | null;
  reportReleased: boolean;
  /** @nullable */
  reportId: number | null;
}

export type ListCasesParams = {
  status?: string;
};

export type GetQuestionnaireStructureParams = {
  mode?: GetQuestionnaireStructureMode;
};

export type GetQuestionnaireStructureMode =
  (typeof GetQuestionnaireStructureMode)[keyof typeof GetQuestionnaireStructureMode];

export const GetQuestionnaireStructureMode = {
  self: "self",
  external: "external",
} as const;