# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Leadership Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (demo mode - select user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /assessments/self/{caseId}:
    get:
      operationId: getSelfAssessment
      tags: [assessments]
      summary: Get self-assessment instance for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Self-assessment instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentInstance"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assessments/token/{token}:
    get:
      operationId: getAssessmentByToken
      tags: [assessments]
      summary: Get assessment by respondent token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assessment instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentInstance"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assessments/{id}/answers:
    get:
      operationId: getAnswers
      tags: [questionnaire]
      summary: Get saved answers for an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Answers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseSet"
    post:
      operationId: saveAnswers
      tags: [questionnaire]
      summary: Save answers for an assessment (partial save)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveAnswersBody"
      responses:
        "200":
          description: Answers saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseSet"
  /assessments/{id}/submit:
    post:
      operationId: submitAssessment
      tags: [questionnaire]
      summary: Submit a completed assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveAnswersBody"
      responses:
        "200":
          description: Assessment submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentInstance"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, participant]
      required:
        - id
        - name
        - email
        - role
    LoginBody:
      type: object
      properties:
        userId:
          type: integer
      required:
        - userId
...
    AssessmentInstance:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: integer
        type:
          type: string
          enum: [self, external]
        status:
          type: string
          enum: [pending, in_progress, submitted]
        respondentName:
          type: ["string", "null"]
        respondentEmail:
          type: ["string", "null"]
        token:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - caseId
        - type
        - status
        - respondentName
        - respondentEmail
        - token
        - submittedAt
        - createdAt
    ResponseSet:
      type: object
      properties:
        assessmentInstanceId:
          type: integer
        answers:
          type: object
          additionalProperties:
            type: ["number", "null"]
        comments:
          type: object
          additionalProperties:
            type: string
        savedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - assessmentInstanceId
        - answers
        - comments
        - savedAt
    SaveAnswersBody:
      type: object
      properties:
        answers:
          type: object
          additionalProperties:
            type: ["number", "null"]
        comments:
          type: object
          additionalProperties:
            type: string
      required:
        - answers
        - comments
...